Added additional message fields
In Object State message children filed is meant to contain the name of the children bodies for the body and the field joint_positions should contain the joint_positions of the attached children.

In object command message, the field position_controller_mask is to enable position control on specific joints based on the mask and effort controller on the rest. The field enable_position_controller if set to 1 overrides the mask and forces all the joints into position control mode